fixed errors

+++ js/task-3.js
@@ -1,6 +1,6 @@
 function getElementWidth(content, padding, border) {
-  const contentValue = parseInt(content);
-  const paddingValue = parseInt(padding);
+  const contentValue = parseFloat(content);
+  const paddingValue = parseFloat(padding);
   const borderValue = parseFloat(border);
   return contentValue + paddingValue * 2 + borderValue * 2;
 }
